+++ includes/class-testimonials-widget-categories-widget.php
@@ -20,12 +20,12 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 require_once AIHR_DIR_INC . 'class-aihrus-widget.php';
 
-if ( class_exists( 'Testimonials_Widget_Widget_Categories' ) )
+if ( class_exists( 'Testimonials_Widget_Categories_Widget' ) )
 	return;
 
 
-class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
-	const ID = 'tw_widget_categories';
+class Testimonials_Widget_Categories_Widget extends Aihrus_Widget {
+	const ID = 'tw_categories_widget';
 
 	public static $title;
 
@@ -62,7 +62,7 @@ class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
 		if ( $d ) {
 			$cat_args['show_option_none'] = esc_html__( 'Select Testimonials Category', 'testimonials-widget' );
 
-			wp_dropdown_categories( apply_filters( 'tw_widget_categories_dropdown_args', $cat_args ) );
+			wp_dropdown_categories( apply_filters( 'tw_categories_widget_dropdown_args', $cat_args ) );
 			?>
 
 			<script type='text/javascript'>
@@ -84,7 +84,7 @@ class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
 			<?php
 				$cat_args['title_li'] = '';
 
-				wp_list_categories( apply_filters( 'tw_widget_categories_args', $cat_args ) );
+				wp_list_categories( apply_filters( 'tw_categories_widget_args', $cat_args ) );
 			?>
 			</ul>
 			<?php
@@ -125,7 +125,7 @@ class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
 			$form_parts[ $id ] = wp_parse_args( $parts, self::$default );
 		}
 
-		$form_parts = apply_filters( 'tw_widget_categories_options', $form_parts );
+		$form_parts = apply_filters( 'tw_categories_widget_options', $form_parts );
 
 		return $form_parts;
 	}